refactor: split subprocess runtime into dedicated microcrate
