+++ crates/perl-lsp-tooling/Cargo.toml
@@ -4,7 +4,7 @@ version.workspace = true
 edition.workspace = true
 rust-version.workspace = true
 authors.workspace = true
-description = "Tooling integration for Perl LSP (perltidy, perlcritic, subprocess abstraction)"
+description = "Tooling integration for Perl LSP (perltidy and perlcritic)"
 readme = "README.md"
 license = "MIT OR Apache-2.0"
 repository.workspace = true
@@ -21,7 +21,7 @@ default = ["lsp-compat"]
 lsp-compat = ["dep:lsp-types"]
 
 [dependencies]
-perl-tdd-support = { workspace = true }
+perl-subprocess-runtime = { workspace = true }
 perl-parser-core = { workspace = true }
 moka = { version = "0.12", features = ["sync"] }
 lsp-types = { version = "0.97.0", optional = true }
@@ -29,6 +29,7 @@ serde = { version = "1.0.228", features = ["derive"] }
 
 [dev-dependencies]
 perl-tdd-support = { workspace = true }
+perl-subprocess-runtime = { workspace = true, features = ["mock"] }
 criterion = "0.8"
 
 [[bench]]
